PRLC100-221 - Linking screening questions in c100 rebuild
