+++ src/main/steps/c100-rebuild/c100sequence.ts
@@ -93,9 +93,9 @@ import {
 
   /** Screening Questions */
   C100_SCREENING_QUESTIONS_CONSENT_AGREEMENT,
-  C100_SCREENING_QUESTIONS_ALTERNATIVE_SOLUTIONS,
+  C100_SCREENING_QUESTIONS_ALTERNATIVE_RESOLUTION,
   C100_SCREENING_QUESTIONS_LEGAL_RESPRESENTATION,
-  C100_LEGAL_REPRESENTATION_APPLICATION,
+  C100_SCREENING_QUESTIONS_LEGAL_REPRESENTATION_APPLICATION,
   C100_SCREENING_QUESTIONS_PERMISSIONS_REQUEST,
   C100_SCREENING_QUESTIONS_PERMISSIONS_WHY,
   C100_SCREENING_QUESTIONS_COURT_PERMISSION,
@@ -389,7 +389,7 @@ export const C100Sequence: Step[] = [
   {
     url: C100_CHILD_ADDRESS,
     showInSection: Sections.C100,
-    getNextStep: () => C100_CHILD_ADDRESS,
+    getNextStep: () => C100_SCREENING_QUESTIONS_CONSENT_AGREEMENT,
   },
   {
     url: C100_DOCUMENT_SUBMISSION,
@@ -584,33 +584,36 @@ export const C100Sequence: Step[] = [
   {
     url: C100_SCREENING_QUESTIONS_CONSENT_AGREEMENT,
     showInSection: Sections.C100,
-    getNextStep: () => C100_SCREENING_QUESTIONS_CONSENT_AGREEMENT,
+    getNextStep: (data: Partial<Case>) =>
+      data.sq_writtenAgreement === YesOrNo.YES
+        ? C100_TYPE_ORDER_SELECT_COURT_ORDER
+        : C100_SCREENING_QUESTIONS_ALTERNATIVE_RESOLUTION,
   },
   {
-    url: C100_SCREENING_QUESTIONS_ALTERNATIVE_SOLUTIONS,
+    url: C100_SCREENING_QUESTIONS_ALTERNATIVE_RESOLUTION,
     showInSection: Sections.C100,
-    getNextStep: () => C100_SCREENING_QUESTIONS_ALTERNATIVE_SOLUTIONS,
+    getNextStep: () => C100_SCREENING_QUESTIONS_LEGAL_RESPRESENTATION,
   },
   {
     url: C100_SCREENING_QUESTIONS_LEGAL_RESPRESENTATION,
     showInSection: Sections.C100,
     getNextStep: (data: Partial<Case>) =>
       data.sq_legalRepresentation === YesOrNo.YES
-        ? C100_SCREENING_QUESTIONS_LEGAL_RESPRESENTATION
-        : C100_SCREENING_QUESTIONS_LEGAL_RESPRESENTATION,
+        ? C100_SCREENING_QUESTIONS_LEGAL_REPRESENTATION_APPLICATION
+        : C100_SCREENING_QUESTIONS_COURT_PERMISSION,
   },
   {
-    url: C100_LEGAL_REPRESENTATION_APPLICATION,
+    url: C100_SCREENING_QUESTIONS_LEGAL_REPRESENTATION_APPLICATION,
     showInSection: Sections.C100,
     getNextStep: data =>
       data.sq_legalRepresentationApplication === YesOrNo.YES
-        ? C100_LEGAL_REPRESENTATION_APPLICATION
-        : C100_CONFIDENTIALITY_START,
+        ? C100_SCREENING_QUESTIONS_LEGAL_REPRESENTATION_APPLICATION
+        : C100_SCREENING_QUESTIONS_COURT_PERMISSION,
   },
   {
     url: C100_SCREENING_QUESTIONS_PERMISSIONS_REQUEST,
     showInSection: Sections.C100,
-    getNextStep: () => C100_SCREENING_QUESTIONS_PERMISSIONS_REQUEST,
+    getNextStep: () => C100_MIAM_OTHER_PROCEEDINGS,
   },
   {
     url: C100_SCREENING_QUESTIONS_ALTERNATIVE_ROUTES,
@@ -620,7 +623,7 @@ export const C100Sequence: Step[] = [
   {
     url: C100_SCREENING_QUESTIONS_PERMISSIONS_WHY,
     showInSection: Sections.C100,
-    getNextStep: () => C100_SCREENING_QUESTIONS_PERMISSIONS_WHY,
+    getNextStep: () => C100_SCREENING_QUESTIONS_PERMISSIONS_REQUEST,
   },
   {
     url: C100_C1A_CHILD_ABDUCTION_THREATS,
@@ -640,7 +643,7 @@ export const C100Sequence: Step[] = [
     showInSection: Sections.C100,
     getNextStep: (data: Partial<Case>) =>
       data.sq_courtPermissionRequired === YesOrNo.YES
-        ? C100_SCREENING_QUESTIONS_COURT_PERMISSION
-        : C100_SCREENING_QUESTIONS_COURT_PERMISSION,
+        ? C100_SCREENING_QUESTIONS_PERMISSIONS_WHY
+        : C100_MIAM_OTHER_PROCEEDINGS,
   },
 ];
