+++ src/main/steps/c100-rebuild/screening-questions/alternative-resolution/content.test.ts
@@ -1,7 +1,10 @@
-import { TranslationFn } from '../../../../../app/controller/GetController';
-import { FormContent } from '../../../../../app/form/Form';
+import languageAssertions from '../../../../../test/unit/utils/languageAssertions';
+import { FormContent, LanguageLookup } from '../../../../app/form/Form';
+import { CommonContent, generatePageContent } from '../../../common/common.content';
 
-const en = () => ({
+import { generateContent } from './content';
+
+const en = {
   title: 'Before you go to court',
   titleDetail:
     'Before you continue with the application, think about what the court process involves and whether you need to go to court.',
@@ -37,9 +40,9 @@ const en = () => ({
   para11: 'If you do not have a legal representative, you can find information about',
   para11Link:
     '<a href="https://www.gov.uk/represent-yourself-in-court" class="govuk-link" rel="external" target="_blank">how to represent yourself in court</a>',
-});
+};
 
-const cy = () => ({
+const cy = {
   title: 'Before you go to court - welsh',
   titleDetail:
     'Before you continue with the application, think about what the court process involves and whether you need to go to court. - welsh',
@@ -75,27 +78,30 @@ const cy = () => ({
   para11: 'If you do not have a legal representative, you can find information about',
   para11Link:
     '<a href="https://www.gov.uk/represent-yourself-in-court" class="govuk-link" rel="external" target="_blank">how to represent yourself in court</a>',
+};
+
+describe('miam->have document signed by mediator or not', () => {
+  const commonContent = { language: 'en', userCase: { applyingWith: 'alone' } } as unknown as CommonContent;
+  const generatedContent = generateContent(commonContent) as Record<string, never>;
+  const form = generatedContent.form as FormContent;
+  // eslint-disable-next-line jest/expect-expect
+  test('should return correct english content', () => {
+    languageAssertions('en', en, () => generateContent(commonContent));
+  });
+
+  // eslint-disable-next-line jest/expect-expect
+  test('should return correct welsh content', () => {
+    languageAssertions('cy', cy, () => generateContent({ ...commonContent, language: 'cy' }));
+  });
+
+  test('should contain Continue button', () => {
+    expect(
+      (form?.submit?.text as LanguageLookup)(generatePageContent({ language: 'en' }) as Record<string, never>)
+    ).toBe('Continue');
+  });
+  test('should contain SaveAndComeLater button', () => {
+    expect(
+      (form?.saveAndComeLater?.text as LanguageLookup)(generatePageContent({ language: 'en' }) as Record<string, never>)
+    ).toBe('Save and come back later');
+  });
 });
-
-const languages = {
-  en,
-  cy,
-};
-
-export const form: FormContent = {
-  fields: {},
-  submit: {
-    text: l => l.onlycontinue,
-  },
-  saveAndComeLater: {
-    text: l => l.saveAndComeLater,
-  },
-};
-
-export const generateContent: TranslationFn = content => {
-  const translations = languages[content.language]();
-  return {
-    ...translations,
-    form,
-  };
-};
