+++ src/main/steps/c100-rebuild/screening-questions/alternative-resolution/template.njk
